+++ application/models/FlightEntity.php
@@ -8,40 +8,32 @@
     class FlightEntity extends Entity {
 
         protected $uniqueid;
-        protected $departureLocation;
-        protected $destinationLocation;
-        protected $departureLocationC;
-        protected $destinationLocationC;
+        protected $departureAirport;
+        protected $destinationAirport;
         protected $departureTime;
         protected $arrivalTime;
         protected $aircraftCode;
 
-        function __construct ($uniqueid = "", $departureLocation = "", $destinationLocation = "", $departureLocationC = "", $destinationLocationC = "", $departureTime = "", $arrivalTime = "", $aircraftCode = "") {
+        function __construct ($uniqueid = null, $destinationAirport = null, $departureAirport = null, $departureTime = null, $arrivalTime = null, $aircraftCode = null) {
             parent::__construct();
             $this -> uniqueid = $uniqueid;
-            $this -> departureLocation = $departureLocation;
-            $this -> destinationLocation = $destinationLocation;
-            $this -> departureLocationC = $departureLocationC;
-            $this -> destinationLocationC = $destinationLocationC;
+            $this -> destinationAirport = $destinationAirport;
+            $this -> departureAirport = $departureAirport;
             $this -> departureTime = $departureTime;
             $this -> arrivalTime = $arrivalTime;
             $this -> aircraftCode = $aircraftCode;
         }
 
-        public function setDepartureLocation($value) {
-            $this -> departureLocation = $value;
+        public function setUniqueid ($value) {
+            $this -> uniqueid = $value;
         }
 
-        public function setDestinationLocation ($value) {
-            $this -> destinationLocation = $value;
+        public function setDepartureAirport ($value) {
+            $this -> departureAirport = $value;
         }
 
-        public function setDepartureLocationC ($value) {
-            $this -> departureLocationC = $value;
-        }
-
-        public function setDestinationLocationC ($value) {
-            $this -> destinationLocationC = $value;
+        public function setDestinationAirport($value) {
+            $this -> destinationAirport = $value;
         }
 
         public function setDepartureTime ($value) {
@@ -56,20 +48,16 @@
             $this -> aircraftCode = $value;
         }
 
-        public function getDepartureLocation() {
-            return $this -> departureLocation;
+        public function getUniqueid () {
+            return $this -> uniqueid;
+        }
+
+        public function getDepartureAirport() {
+            return $this -> departureAirport;
         }
 
         public function getDestinationLocation () {
-            return $this -> destinationLocation;
-        }
-
-        public function getDepartureLocationC () {
-            return $this -> departureLocationC;
-        }
-
-        public function getDestinationLocationC () {
-            return $this -> destinationLocationC;
+            return $this -> destinationAirport;
         }
 
         public function getDepartureTime () {
@@ -84,12 +72,20 @@
             return $this -> aircraftCode;
         }
 
+        public static function create_flight_from_obj_with_airport ($object, $destAirport, $departAirport) {
+            return new FlightEntity ($object.id, $destAirport, $departAirport, $object.departureTime, $object.arrivalTime, $object.aircraftCode);
+        }
+
+        public static function create_flight_from_arr_with_airport ($arr, $destAirport, $departAirport) {
+            return new FlightEntity ($arr["uniqueid"], $destAirport, $departAirport, $arr["departureTime"], $arr["arrivalTime"], $arr["aircraftCode"]);
+        }
+
         public static function create_flight_from_obj ($object) {
-            return new FlightEntity ($object.id, $object.departureLocation, $object.destinationLocation, $object.departureLocationC, $object.destinationLocationC, $object.departureTime, $object.arrivalTime, $object.aircraftCode);
+            return new FlightEntity ($object.id, $object.destinationAirport, $object.departureAirport, $object.departureTime, $object.arrivalTime, $object.aircraftCode);
         }
 
         public static function create_flight_from_arr ($arr) {
-            return new FlightEntity ($arr["uniqueid"], $arr["departureLocation"], $arr["destinationLocation"], $arr["departureLocationC"], $arr["destinationLocationC"], $arr["departureTime"], $arr["arrivalTime"], $arr["aircraftCode"]);
+            return new FlightEntity ($arr["uniqueid"], $arr["destinationAirport"], $arr["departureAirport"], $arr["departureTime"], $arr["arrivalTime"], $arr["aircraftCode"]);
         }
        
     }