+++ changelog.md
@@ -3,6 +3,12 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
+
+## [Unreleased] - 2017-11-10 - YR
+### Changed
+- Changed flightentity to contain airport objects
+- Modified Airlines to work, had bug that was using wrong data set
+
 ## [Unreleased] - 2017-11-10 - YR
 ### Changed
 - Changed flightentity id to uniqueid